+++ Changelog.md
@@ -5,6 +5,7 @@
 - Tapping an action now triggers a dismiss of the notification. This also fixes the issue where actions would be called multiple times.
 - Chevron image is now having a max size of 20x20.
 - Some small memory fixes
+- Chevron image is now set inside the style, which is more logic.
 
 ### 1.0 (2017-07-14)